friendlist n pendingreq all functional
