--- src/app/user/[username]/pending-requests/page.jsx
@@ -1,10 +1,0 @@
-import PendingRequest from "@/components/PendingRequest";
-import React from "react";
-
-export default function page({ params }) {
-  return (
-    <div>
-      <PendingRequest username={params.username} />
-    </div>
-  );
-}
